inn-crawler/requirements.txt
Flask~=2.2.2
requests~=2.22.0
Cerberus~=1.3.4
gunicorn~=20.1.0
psycopg2-binary~=2.9.5
python-dotenv~=1.0.0
flask-crontab~=0.1.2
